Refactors classes into their respective files. Fixes quitting Redis client upon exiting applciation
